Document_qna_Chatbot/requirements.txt
streamlit
langchain
langchain-groq
langchain-google-genai
langchain-core
langchain-community
pypdf
faiss-cpu
python-dotenv
SQLAlchemy
